+++ tf-modules/ingest/files-to-granules-task.tf
@@ -5,8 +5,8 @@ resource "aws_lambda_function" "files_to_granules_task" {
   handler          = "index.handler"
   role             = var.lambda_processing_role_arn
   runtime          = "nodejs16.x"
-  timeout          = lookup(var.lambda_timeouts, "files_to_granules_task_timeout", 300)
-  memory_size      = lookup(var.lambda_memory_sizes, "files_to_granules_task_memory_size", 1024)
+  timeout          = lookup(var.lambda_timeouts, "FilesToGranules", 300)
+  memory_size      = lookup(var.lambda_memory_sizes, "FilesToGranules", 1024)
 
   layers = [var.cumulus_message_adapter_lambda_layer_version_arn]