+++ tf-modules/ingest/update-cmr-access-constraints.tf
@@ -6,8 +6,8 @@ resource "aws_lambda_function" "update_cmr_access_constraints_task" {
   handler          = "index.handler"
   role             = var.lambda_processing_role_arn
   runtime          = "nodejs16.x"
-  timeout          = lookup(var.lambda_timeouts, "update_cmr_access_constraints_task_timeout", 300)
-  memory_size      = lookup(var.lambda_memory_sizes, "update_cmr_access_constraints_task_memory_size", 512)
+  timeout          = lookup(var.lambda_timeouts, "UpdateCmrAccessConstraints", 300)
+  memory_size      = lookup(var.lambda_memory_sizes, "UpdateCmrAccessConstraints", 512)
 
   layers = [var.cumulus_message_adapter_lambda_layer_version_arn]