+++ tf-modules/s3-replicator/variables.tf
@@ -43,3 +43,15 @@ variable "source_bucket" {
 variable "source_prefix" {
   type = string
 }
+
+variable "lambda_timeouts" {
+  description = "Configurable map of timeouts for lambdas"
+  type = map(number)
+  default = {}
+}
+
+variable "lambda_memory_sizes" {
+  description = "Configurable map of memory sizes for lambdas"
+  type = map(number)
+  default = {}
+}